% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter-package.R
\docType{data}
\name{aaAlignment}
\alias{aaAlignment}
\title{Example protein alignment.}
\format{
An object of class \code{AAbin} of length 2907.
}
\usage{
data(afExample)
}
\description{
Example protein alignment.
}
\keyword{datasets}
