% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter-package.R
\docType{data}
\name{afModel}
\alias{afModel}
\title{Example AliFilter model.}
\format{
An object of class \code{alifilter_model} of length 2.
}
\usage{
data(afExample)
}
\description{
Example AliFilter model.
}
\keyword{datasets}
