% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter-package.R
\docType{package}
\name{alifilter-package}
\alias{alifilter}
\alias{alifilter-package}
\title{alifilter: Alignment Filtering Using Machine Learning}
\description{
This package contains functions to use pre-trained AliFilter models in R.
}
\keyword{internal}
