% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter.R
\name{filterAlignment}
\alias{filterAlignment}
\title{Filter an alignment (\code{DNAbin} or \code{AAbin}) using the specified AliFilter
model.}
\usage{
filterAlignment(alignment, model)
}
\arguments{
\item{alignment}{A sequence alignment as an object of class \code{AAbin} or \code{DNAbin} (from
package ape).}

\item{model}{An object of class \code{alifilter_model}, specifying the AliFilter model.}
}
\value{
An alignment (\code{DNAbin} or \code{AAbin}) containing only the columns that have
been preserved according to the AliFilter model.
}
\description{
Filter an alignment (\code{DNAbin} or \code{AAbin}) using the specified AliFilter
model.
}
\examples{
  data(afExample)
  filterAlignment(aaAlignment, afModel)

}
