% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter.R
\name{getAlignmentFeatures}
\alias{getAlignmentFeatures}
\title{Computes the alignment features for a DNA or protein alignment (\code{AAbin} or
\code{DNAbin}).}
\usage{
getAlignmentFeatures(alignment)
}
\arguments{
\item{alignment}{A sequence alignment as an object of class \code{AAbin} or \code{DNAbin} (from
package ape).}
}
\value{
A two-dimensional list of mode numeric containing the features computed for
each alignment column.
}
\description{
Computes the alignment features for a DNA or protein alignment (\code{AAbin} or
\code{DNAbin}).
}
\examples{
  data(afExample)
  features <- getAlignmentFeatures(aaAlignment)

}
