% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter.R
\name{getColumnScores}
\alias{getColumnScores}
\title{Compute column preservation scores, given pre-computed alignment features and
an AliFilter model.}
\usage{
getColumnScores(alignmentFeatures, model)
}
\arguments{
\item{alignmentFeatures}{The pre-computed alignment features (from the \code{getAlignmentFeatures}
function)}

\item{model}{An object of class \code{alifilter_model}, representing the model to use to
compute the column preservation scores.}
}
\value{
A vector of mode numeric containing the column preservation scores (from
0 to 1).
}
\description{
Compute column preservation scores, given pre-computed alignment features and
an AliFilter model.
}
\examples{
  data(afExample)
  features <- getAlignmentFeatures(aaAlignment)
  getColumnScores(features, afModel)

}
