% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter.R
\name{getMask}
\alias{getMask}
\title{Get an alignment mask from an alignment (\code{AAbin} or \code{DNAbin}) using the
specified model.}
\usage{
getMask(alignment, model)
}
\arguments{
\item{alignment}{A sequence alignment as an object of class \code{AAbin} or \code{DNAbin} (from
package ape).}

\item{model}{An object of class \code{alifilter_model}, specifying the AliFilter model.}
}
\value{
A string containing \code{1} for columns that should be preserved and \code{0} for
columns that should be deleted.
}
\description{
Get an alignment mask from an alignment (\code{AAbin} or \code{DNAbin}) using the
specified model.
}
\examples{
  data(afExample)
  getMask(aaAlignment, afModel)

}
