% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter.R
\name{getMaskFromAlignmentFeatures}
\alias{getMaskFromAlignmentFeatures}
\title{Get an alignment mask from pre-computed alignment features.}
\usage{
getMaskFromAlignmentFeatures(alignmentFeatures, model)
}
\arguments{
\item{alignmentFeatures}{The pre-computed alignment features (from the \code{getAlignmentFeatures}
function)}

\item{model}{An object of class \code{alifilter_model}, specifying the AliFilter model.}
}
\value{
A string containing \code{1} for columns that should be preserved and \code{0} for
columns that should be deleted.
}
\description{
Get an alignment mask from pre-computed alignment features.
}
\examples{
  data(afExample)
  features <- getAlignmentFeatures(aaAlignment)
  getMaskFromAlignmentFeatures(features, afModel)

}
