% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alifilter.R
\name{getMaskFromColumnScores}
\alias{getMaskFromColumnScores}
\title{Get an alignment mask from pre-computed column scores.}
\usage{
getMaskFromColumnScores(scores, model)
}
\arguments{
\item{scores}{The pre-computed column preservation scores (from the \code{getColumnScores}
function)}

\item{model}{An object of class \code{alifilter_model}, containing the preservation threshold
used to convert the column scores into the mask.}
}
\value{
A string containing \code{1} for columns that should be preserved and \code{0} for
columns that should be deleted.
}
\description{
Get an alignment mask from pre-computed column scores.
}
\examples{
  data(afExample)
  features <- getAlignmentFeatures(aaAlignment)
  scores <- getColumnScores(features, afModel)
  getMaskFromColumnScores(scores, afModel)

}
